move myVidoeCourses components to components folder
