+++ __tests__/integration/videoPlayer.test.js
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import { ThemeProvider } from "styled-components";
-import VideoPlayer from "../../src/components/videoPlayer";
+import VideoPlayer from "../../src/components/videoCourses/videoPlayer";
 
 const theme = {
   colors: {
